arqui3/services/s4/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
strawberry-graphql==0.217.0
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
reportlab==4.0.9
pandas==2.1.4
openpyxl==3.1.2 